train/requirements.txt
torch
torchvision
pyyaml
pillow
numpy
opencv-python
